Make sed more robust when parse nearest mirror
It looks like the maintainers might have changed the HTML structure of the page so this new sed regex will look for the first occurrence of a string beginning with http:// or https:// instead of just the first string as it was before. Without this an HTML hash anchor is returned instead breaking the download path used in the next step of the provisioning process.